modules/EventProducerAPI/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Flask==2.0.2
gevent==21.8.0
greenlet==1.1.2
grpcio==1.42.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kafka-python==2.0.2
MarkupSafe==2.0.1
protobuf==3.19.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
